requirements.txt
fastapi 
uvicorn 
langchain 
langgraph 
chromadb 
pypdf 
python-dotenv
langchain_community
langchain-openai
langchain-chroma
langchain-huggingface
sentence-transformers